possible improved version of the code
