+++ requirements.txt
@@ -1,2 +1,2 @@
-importlib-resources==5.8
-Shapely~=1.8.2
+importlib-resources==5.12.0
+Shapely==2.0.1
